+++ client/src/views/Repos.vue
@@ -1,5 +1,5 @@
 <template>
-  <section v-if="isAdmin" class="section">
+  <section v-if="profile?.options?.showSettings ?? isAdmin" class="section">
     <BulmaQuickView class="quickview" v-if="showOutput" title="Last execution result" footer="" @close="showOutput=false">
         <p class="is-family-code" v-html="output.split('\n').join('<br>')"></p>
     </BulmaQuickView>        
@@ -29,8 +29,8 @@
               </div>              
               <BulmaAdminTable
                 :dataList="repositoryList"
-                :labels="['Name','Status','Head','Type']"
-                :columns="['name','status','head','icon']"
+                :labels="['Name','Branch','Status','Head','Type']"
+                :columns="['name','branch','status','head','icon']"
                 :filters="['name','status']"
                 :icons="[false,false,false,true]"
                 identifier="name"
@@ -44,8 +44,9 @@
             </div>
             <transition name="add-column" appear>
               <div class="column" v-if="repositoryItem!==undefined && !showDelete">
-                <BulmaInput icon="heading" v-model="repository.name" label="Name" placeholder="my_repo_name" :readonly="repositoryItem!==-1" :required="true" :hasError="v$.repository.name.$invalid" :errors="[]" help="Alphanumeric with dash and hyphen" />
-                <BulmaInput icon="user" v-model="repository.user" label="Username" placeholder="my-user" :hasError="v$.repository.user.$invalid" :errors="[]" help="Alphnumeric with hyphen" />
+                <BulmaInput icon="heading" v-model="repository.name" label="Name" placeholder="my_repo_name" :readonly="repositoryItem!==-1" :required="true" :hasError="v$.repository.name.$invalid" :errors="[]" help="Alphanumeric with underscore and hyphen" />
+                <BulmaInput icon="code-branch" v-model="repository.branch" label="Branch" placeholder="master" :hasError="v$.repository.branch.$invalid" :errors="[]" />
+                <BulmaInput icon="user" v-model="repository.user" label="Username" placeholder="my-user" :hasError="v$.repository.user.$invalid" :errors="[]" help="Alphanumeric with hyphen" />
                 <BulmaInput icon="lock" v-model="repository.password" type="password" label="Password" placeholder="Password or Token" />
                 <BulmaInput :icon="['fab','git']" v-model="repository.uri" label="Uri" placeholder="https://github.com/account/repo.git" :required="true" :hasError="v$.repository.uri.$invalid" :errors="[]" help="Only ssh or https uri's are allowed" />
                 <BulmaInput icon="stopwatch" v-model="repository.cron" label="Cron Schedule" placeholder="*/5 * L * 1,3L" :hasError="v$.repository.cron.$invalid" :errors="[]" help="Minute Hour DayOfMonth Month DayOfWeek" />
@@ -78,11 +79,13 @@
   import { useVuelidate } from '@vuelidate/core'
   import { required, helpers } from '@vuelidate/validators'
 
+
   export default{
     name:"AfRepositories",
     props:{
       authenticated:{type:Boolean},
-      isAdmin:{type:Boolean}
+      isAdmin:{type:Boolean},
+      profile:{type:Object}
     },
     components:{BulmaButton,BulmaInput,BulmaModal,BulmaQuickView,BulmaAdminTable,BulmaCheckbox,BulmaSettingsMenu},
     setup(){
@@ -92,6 +95,7 @@
       return  {
           repository:{
             name:"",
+            branch:"",
             user:"",
             password:"",
             uri:"",
@@ -263,8 +267,14 @@
           required,
           regex : helpers.withParams(
               {description: "User must be a valid repository name",type:"regex"},
-              (value) => !helpers.req(value) || (new RegExp("^[a-z0-9_-]{1,50}$").test(value)) // eslint-disable-line
+              (value) => !helpers.req(value) || (new RegExp("^[a-zA-Z0-9_-]{1,50}$").test(value)) // eslint-disable-line
           )             
+        },
+        branch: {
+          regex : helpers.withParams(
+              {description: "Must be valid git branch name",type:"regex"},
+              (value) => !helpers.req(value) || (new RegExp("^[^\s]+$").test(value)) // eslint-disable-line
+          )     
         },
         uri: {
           required,
@@ -278,7 +288,7 @@
         },
         cron: {
           regex : helpers.withParams(
-              {description: "User must be a valid github user (alphanumeric / hyphens)",type:"regex"},
+              {description: "Cron must be valid format",type:"regex"},
               (value) => !helpers.req(value) || (new RegExp("^[0-9-,*/]+ [0-9-,*/]+ [0-9-,*/L]+ [0-9-,*/]+ [0-9-,*/L]+$").test(value)) // eslint-disable-line
           )                    
         },
@@ -286,7 +296,7 @@
           maxLength:39,
           regex : helpers.withParams(
               {description: "User must be a valid github user (alphanumeric / hyphens)",type:"regex"},
-              (value) => !helpers.req(value) || (new RegExp("^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$").test(value)) // eslint-disable-line
+              (value) => !helpers.req(value) || (new RegExp("^[a-zA-Z0-9-](?:[a-zA-Z0-9-]|-(?=[a-zA-Z0-9-])){0,38}$").test(value)) // eslint-disable-line
           )          
         }
       }