+++ server/src/db/create_settings_table.sql
@@ -9,4 +9,4 @@ CREATE TABLE `settings` (
   `url` varchar(250) DEFAULT NULL,
   `forms_yaml` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL 
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
-INSERT INTO AnsibleForms.settings(mail_server,mail_port,mail_secure,mail_username,mail_password,mail_from,url,forms_yaml) VALUES('',25,0,'','','','','');
+