+++ server/src/lib/common.js
@@ -68,16 +68,51 @@ Helpers.checkCertificate=function(cert){
 }
 // a middleware in the routes to check if use is administrator
 Helpers.checkAdminMiddleware = (req, res, next) =>  {
-      try{
-        if(!req.user.user.roles.includes("admin")) {
-          res.status(401).json(new restResult("error","you are not an admin"))
-        } else {
-          //logger.debug("You are admin, access to user management")
-          next()
-        }
-      }catch(e){
-        res.status(401).json(new restResult("error","you are not an admin"))
-      }
+  try{
+    if(!req.user.user.roles.includes("admin")) {
+      res.status(401).json(new restResult("error","No access",null,"You are not an admin"))
+    } else {
+      //logger.debug("You are admin, access to user management")
+      next()
+    }
+  }catch(e){
+    res.status(401).json(new restResult("error","No access",null,"You are not an admin"))
+  }
+}
+Helpers.checkSettingsMiddleware = (req, res, next) => {
+  try {
+    if (!(req.user.user.options?.showSettings ?? req.user.user.roles.includes("admin"))) {
+      res.status(401).json(new restResult("error", "No access",null,"You do not have access to settings"));
+    } else {
+      next();
+    }
+  } catch (e) {
+    res.status(401).json(new restResult("error", "No access",null,"You do not have access to settings"));
+  }
+}
+
+Helpers.checkDesignerMiddleware = (req, res, next) => {
+  try {
+    if (!(req.user.user.options?.showDesigner ?? req.user.user.roles.includes("admin"))) {
+      res.status(401).json(new restResult("error", "No access",null,"You do not have access to designer"));
+    } else {
+      next();
+    }
+  } catch (e) {
+    res.status(401).json(new restResult("error", "No access",null,"You do not have access to designer"));
+  }
+}
+
+Helpers.checkLogsMiddleware = (req, res, next) => {
+  try {
+    if (!(req.user.user.options?.showLogs ?? req.user.user.roles.includes("admin"))) {
+      res.status(401).json(new restResult("error", "No access",null,"You do not have access to logs"));
+    } else {
+      next();
+    }
+  } catch (e) {
+    res.status(401).json(new restResult("error", "No access",null,"You do not have access to logs"));
+  }
 }
 // checks for passwords from credentials and masks them
 Helpers.logSafe = (v)=>{