+++ server/src/routes/config.routes.js
@@ -1,14 +1,15 @@
 const express = require('express')
 const router = express.Router()
 const configController =   require('../controllers/config.controller');
-const checkAdminMiddleware = require('../lib/common.js').checkAdminMiddleware;
+const checkDesignerMiddleware = require('../lib/common.js').checkDesignerMiddleware;
+const checkSettingsMiddleware = require('../lib/common.js').checkSettingsMiddleware;
 
 // designer routes only (admin !!)
-router.post('/', checkAdminMiddleware, configController.save);
-router.post('/check', checkAdminMiddleware, configController.validate);
-router.post('/restore/:backupName', checkAdminMiddleware, configController.restore);
-router.get('/backups', checkAdminMiddleware, configController.backups);
-router.get('/env', checkAdminMiddleware, configController.env);
+router.post('/', checkDesignerMiddleware, configController.save);
+router.post('/check', checkDesignerMiddleware, configController.validate);
+router.post('/restore/:backupName', checkDesignerMiddleware, configController.restore);
+router.get('/backups', checkDesignerMiddleware, configController.backups);
+router.get('/env', checkSettingsMiddleware, configController.env);
 
 // get the config (no admin)
 router.get('/', configController.findAll);